requirements.txt
Pillow==8.2.0
PySimpleGUI==4.40.0
matplotlib==3.3.4
numpy==1.20.1
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
scikit-image==0.18.1
unittest2==1.1.0
